Move filtering to common filter
